Some fixes and Designchanges

+++ app/src/main/java/de/ntbit/projectearlybird/manager/ModuleChecklistManager.kt
@@ -106,7 +106,6 @@ class ModuleChecklistManager {
         }
     }
 
-    /*
     private fun listenForDeleteChecklistItem() {
         val parseQuery = ParseQuery.getQuery(ModuleChecklistItem::class.java)
         val subscriptionHandling: SubscriptionHandling<ModuleChecklistItem> =
@@ -114,15 +113,10 @@ class ModuleChecklistManager {
         subscriptionHandling.handleEvent(SubscriptionHandling.Event.DELETE) {_, item ->
             val handler = Handler(Looper.getMainLooper())
             handler.post {
-                Log.d("CUSTOMDEBUG", "$simpleClassName - " +
-                        "CurrentUser: ${mUserManager.getCurrentUser().objectId}, " +
-                        "got update on item:\n$item")
-                processUpdateOnChecklistItem(item)
+                deleteChecklistItem(ChecklistItem(item), false)
             }
         }
     }
-    */
-
 
     private fun processUpdateOnChecklistItem(item: ModuleChecklistItem) {
         val group = item.associatedModule.associatedGroup
@@ -154,7 +148,7 @@ class ModuleChecklistManager {
         item.saveEventually()
     }
 
-    fun deleteChecklistItem(checklistItem: ChecklistItem) {
+    fun deleteChecklistItem(checklistItem: ChecklistItem, deleteFromDatabase: Boolean) {
         val item = checklistItem.getModuleChecklistItem()
         val group = item.associatedModule.associatedGroup
         val position = adapterMap[group]!!.getAdapterPosition(checklistItem)
@@ -162,9 +156,7 @@ class ModuleChecklistManager {
         adapterMap[group]!!.notifyItemRangeChanged(position, adapterMap[group]!!.itemCount -1)
         //adapterMap[group]!!.notifyItemRemoved(position)
         checklistItemMap[group]!!.remove(item)
-        deleteItemOnDatabase(item)
-        Log.d("CUSTOMDEBUG", "$simpleClassName - Count: ${adapterMap[group]!!.itemCount} " +
-                "after deleting on Position $position")
+        if(deleteFromDatabase) deleteItemOnDatabase(item)
     }
 
     private fun deleteItemOnDatabase(item: ModuleChecklistItem) {
